refactor(test): migrate connection tests to feature directory
- Move connection tests from tests/connection/ to tests/feature/connection/
  to align with project test structure conventions
- Add session-aware connection pool test fixtures:
  - BackendPool and AsyncBackendPool fixtures
  - PoolConfig for pool configuration testing
  - SessionAwareMixin and AsyncSessionAwareMixin test models
  - pool_with_tables fixture for pool table initialization
